security: fix IDOR server actions, SSRF gaps, and hardcoded webhook
- Relocate computeChangeMap, autoApproveZeroDiffCases, and triggerAIDiffAnalysis
  out of "use server" files into plain lib modules so they are no longer
  RPC-invocable. They read/mutate another team's build data and spend the
  owning team's GitHub/AI budget; trusted callers (build finalization,
  ownership-checked routes, guarded setBuildManualScope) import them directly.
- Add safeOutboundFetch (validates URL + re-validates every redirect hop) and
  apply the SSRF guard to "Analyze URL" (was redirect:"follow", unguarded).
- Guard notification webhook senders (custom/Slack/Discord) and the webhook
  tester with assertSafeOutboundUrl; switch to redirect:"manual" and cap
  reflected response bodies so the tester is no longer an SSRF read oracle.
- Remove the hardcoded Discord webhook URL+token fallback from public-shares;
  the internal share ping is now env-only (opt-in).
